+++ src/main/java/org/programmers/signalbuddyfinal/domain/feedback/controller/FeedbackController.java
@@ -1,7 +1,6 @@
 package org.programmers.signalbuddyfinal.domain.feedback.controller;
 
 import jakarta.validation.Valid;
-import java.time.LocalDate;
 import lombok.RequiredArgsConstructor;
 import org.programmers.signalbuddyfinal.domain.feedback.dto.FeedbackRequest;
 import org.programmers.signalbuddyfinal.domain.feedback.dto.FeedbackResponse;
@@ -12,10 +11,7 @@ import org.programmers.signalbuddyfinal.global.dto.CustomUser2Member;
 import org.programmers.signalbuddyfinal.global.dto.PageResponse;
 import org.programmers.signalbuddyfinal.global.response.ApiResponse;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort.Direction;
 import org.springframework.data.web.PageableDefault;
-import org.springframework.format.annotation.DateTimeFormat;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -61,30 +57,6 @@ public class FeedbackController {
             ApiResponse.createSuccess(
                 feedbackService.searchFeedbackList(
                     pageable, request, crossroadId
-                )
-            )
-        );
-    }
-
-    @GetMapping(headers = HttpHeaders.AUTHORIZATION)
-    public ResponseEntity<ApiResponse<PageResponse<FeedbackResponse>>> searchFeedbackListByAdmin(
-        @PageableDefault(sort = {"createdAt"}, direction = Direction.DESC)
-        Pageable pageable,
-        @ModelAttribute FeedbackSearchRequest request,
-        @RequestParam(value = "deleted", required = false)
-        Boolean deleted,
-        @RequestParam(name = "start-date", required = false)
-        @DateTimeFormat(pattern = "yyyy-MM-dd")
-        LocalDate startDate,
-        @RequestParam(name = "end-date", required = false)
-        @DateTimeFormat(pattern = "yyyy-MM-dd")
-        LocalDate endDate,
-        @CurrentUser CustomUser2Member user
-    ) {
-        return ResponseEntity.ok(
-            ApiResponse.createSuccess(
-                feedbackService.searchFeedbackListByAdmin(
-                    pageable, request, deleted, startDate, endDate, user
                 )
             )
         );
